+++ spring-integration-ip/src/main/java/org/springframework/integration/ip/tcp/connection/TcpNetClientConnectionFactory.java
@@ -44,59 +44,17 @@ public class TcpNetClientConnectionFactory extends
 		super(host, port);
 	}
 
-	/**
-	 * @throws IOException
-	 * @throws SocketException
-	 * @throws Exception
-	 */
 	@Override
-	protected TcpConnectionSupport obtainConnection() throws Exception {
-		boolean locked = false;
-		if (!this.isSingleUse()) {
-			this.theConnectionLock.readLock().lockInterruptibly();
-			locked = true;
-		}
-		try {
-			TcpConnectionSupport theConnection = this.getTheConnection();
-			if (theConnection != null && theConnection.isOpen()) {
-				return theConnection;
-			}
-		}
-		finally {
-			if (locked) {
-				this.theConnectionLock.readLock().unlock();
-				locked = false;
-			}
-		}
-
-		if (!this.isSingleUse()) {
-			this.theConnectionLock.writeLock().lockInterruptibly();
-			locked = true;
-		}
-		try {
-			if (logger.isDebugEnabled()) {
-				logger.debug("Opening new socket connection to " + this.getHost() + ":" + this.getPort());
-			}
-
-			Socket socket = createSocket(this.getHost(), this.getPort());
-			setSocketAttributes(socket);
-			TcpConnectionSupport connection = new TcpNetConnection(socket, false, this.isLookupHost(),
-					this.getApplicationEventPublisher(), this.getComponentName());
-			connection = wrapConnection(connection);
-			initializeConnection(connection, socket);
-			this.getTaskExecutor().execute(connection);
-			this.harvestClosedConnections();
-			if (!this.isSingleUse()) {
-				this.setTheConnection(connection);
-			}
-			connection.publishConnectionOpenEvent();
-			return connection;
-		}
-		finally {
-			if (locked) {
-				this.theConnectionLock.writeLock().unlock();
-			}
-		}
+	protected TcpConnectionSupport buildNewConnection() throws IOException, SocketException, Exception {
+		Socket socket = createSocket(this.getHost(), this.getPort());
+		setSocketAttributes(socket);
+		TcpConnectionSupport connection = new TcpNetConnection(socket, false, this.isLookupHost(),
+				this.getApplicationEventPublisher(), this.getComponentName());
+		connection = wrapConnection(connection);
+		initializeConnection(connection, socket);
+		this.getTaskExecutor().execute(connection);
+		this.harvestClosedConnections();
+		return connection;
 	}
 
 	@Override
