+++ spring-integration-ip/src/main/java/org/springframework/integration/ip/tcp/connection/TcpNioClientConnectionFactory.java
@@ -18,7 +18,6 @@ package org.springframework.integration.ip.tcp.connection;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
-import java.net.SocketException;
 import java.nio.ByteBuffer;
 import java.nio.channels.CancelledKeyException;
 import java.nio.channels.ClosedChannelException;
@@ -61,13 +60,9 @@ public class TcpNioClientConnectionFactory extends
 		super(host, port);
 	}
 
-	/**
-	 * @throws Exception
-	 * @throws IOException
-	 * @throws SocketException
-	 */
 	@Override
-	protected TcpConnectionSupport obtainConnection() throws Exception {
+	protected void checkActive() throws IOException {
+		super.checkActive();
 		int n = 0;
 		while (this.selector == null) {
 			try {
@@ -76,63 +71,29 @@ public class TcpNioClientConnectionFactory extends
 				Thread.currentThread().interrupt();
 			}
 			if (n++ > 600) {
-				throw new Exception("Factory failed to start");
+				throw new IOException("Factory failed to start");
 			}
 		}
+	}
 
-		boolean locked = false;
-		if (!this.isSingleUse()) {
-			this.theConnectionLock.readLock().lockInterruptibly();
-			locked = true;
+	@Override
+	protected TcpConnectionSupport buildNewConnection() throws Exception {
+		SocketChannel socketChannel = SocketChannel.open(new InetSocketAddress(this.getHost(), this.getPort()));
+		setSocketAttributes(socketChannel.socket());
+		TcpNioConnection connection = this.tcpNioConnectionSupport.createNewConnection(
+				socketChannel, false, this.isLookupHost(), this.getApplicationEventPublisher(), this.getComponentName());
+		connection.setUsingDirectBuffers(this.usingDirectBuffers);
+		connection.setTaskExecutor(this.getTaskExecutor());
+		TcpConnectionSupport wrappedConnection = wrapConnection(connection);
+		initializeConnection(wrappedConnection, socketChannel.socket());
+		socketChannel.configureBlocking(false);
+		if (this.getSoTimeout() > 0) {
+			connection.setLastRead(System.currentTimeMillis());
 		}
-		try {
-			TcpConnectionSupport theConnection = this.getTheConnection();
-			if (theConnection != null && theConnection.isOpen()) {
-				return theConnection;
-			}
-		}
-		finally {
-			if (locked) {
-				this.theConnectionLock.readLock().unlock();
-				locked = false;
-			}
-		}
-
-		if (!this.isSingleUse()) {
-			this.theConnectionLock.writeLock().lockInterruptibly();
-			locked = true;
-		}
-
-		try {
-			if (logger.isDebugEnabled()) {
-				logger.debug("Opening new socket channel connection to " + this.getHost() + ":" + this.getPort());
-			}
-			SocketChannel socketChannel = SocketChannel.open(new InetSocketAddress(this.getHost(), this.getPort()));
-			setSocketAttributes(socketChannel.socket());
-			TcpNioConnection connection = this.tcpNioConnectionSupport.createNewConnection(
-					socketChannel, false, this.isLookupHost(), this.getApplicationEventPublisher(), this.getComponentName());
-			connection.setUsingDirectBuffers(this.usingDirectBuffers);
-			connection.setTaskExecutor(this.getTaskExecutor());
-			TcpConnectionSupport wrappedConnection = wrapConnection(connection);
-			initializeConnection(wrappedConnection, socketChannel.socket());
-			socketChannel.configureBlocking(false);
-			if (this.getSoTimeout() > 0) {
-				connection.setLastRead(System.currentTimeMillis());
-			}
-			this.channelMap.put(socketChannel, connection);
-			newChannels.add(socketChannel);
-			selector.wakeup();
-			if (!this.isSingleUse()) {
-				this.setTheConnection(connection);
-			}
-			connection.publishConnectionOpenEvent();
-			return wrappedConnection;
-		}
-		finally {
-			if (locked) {
-				this.theConnectionLock.writeLock().unlock();
-			}
-		}
+		this.channelMap.put(socketChannel, connection);
+		newChannels.add(socketChannel);
+		selector.wakeup();
+		return wrappedConnection;
 	}
 
 	/**
